setup the router for product and service

+++ router/categoryrouter.py
@@ -5,10 +5,11 @@ from config.database  import get_db
 from schemas.Category import CategoryCreate,CategoryResponse,CategoryBase
 from service.Categoryservice import Categoryservice
 from fastapi import Depends
+from typing import List
 
 
 
-router=APIRouter(prefix="/category",tags=['Category'])
+router=APIRouter(prefix="/category",tags=['category'])
 
 
 
@@ -21,18 +22,27 @@ def createcategory(category:CategoryCreate,db:Session=Depends(get_db)):
     return Categoryservice.createcategory(category,db)
     
 
-@app.get("/getcategory",status_code=status.HTTP_201_CREATED,response_model=List[CategoryResponse])
+@app.get("/getcategory",status_code=status.HTTP_200_OK,response_model=List[CategoryResponse])
 def getallcategory(category:createCategory,db:Session=Depends(get_db)):
     return Categoryservice.getallcategory(db)
 
 @app.delete("/deletecategory/{category_id}")
 def deletecategory(category_id:int, db:Session=Depends(get_db)):
+    if category_id is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
+
     return Categoryservice.deletecategory(category_id,db)
 
 @app.put("/updatecategory/{category_id}")
 def updatedcategory(category_id:int, category:CategoryBase, db:Session=Depends(get_db)):
+    if category or category_id is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
+
     return Categoryservice.updatedcategory(category_id, category, db)
 
 @app.put("/getcategorybyId/{category_id}")
 def getcategorybyId(category_id:int, db:Session=Depends(get_db)):
+    if category is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category not found")
+
     return Categoryservice.getcategorybyId(category_id, db)
